Fix navbar for wider screens
For small screens the position fixed is OK because the menu is
toggled by the button and it takes up all screen until closed.
Otherwise when scrolling down the menu would end up above and
out of the screen (not desired).

But for wider screens the menu links are displayed directly on the
top of the page, but should not be fixed because then when scrolling
they would still be always on top of the screen (not desired).