Reserve space for images before they load
When images load the layout changes because before they are loaded
the space they will occupy has no height, so when they load they
take up the space they need and the rest of the layout shifts position.

To fix this, a CSS trick using an image container with as much padding
as needed to reserve space for the image is used. The image is then
absolutely positioned with respect to the image container, and voila.